ARA-53 Send changed fields on an update jira issue
New 'updatedFields' map that will contain the field name as the key and
the value as a map of 'oldValue' and 'newValue' strings.

While I was at it, I reworked how web hooks are tested using a new fancy
JSON query language and optimized for running tests w/o selenium.

Change-Id: I5d2bed614ef9c59c3c8d03359e95913c9f1136ee