ARA-63 Remove erroneous 'accepts' option from api
RA.request exposed and documented the 'accepts' option, passing it
through to jQuery, but it was an incorrect usage of the jq param to do
so.  The fix is to remove the option from docs and usage on the
host-side of the bridge, and to then override the host-side Accept
header in all requests to always send '*/*'.

Change-Id: Ife00dfa61237d7b21a9dd0eb0534358ae328764d