ARA-62 Rewrite HTTP content retrieval engine
Now use a big-pipe like retrieval engine that allows the Confluence page to be returned
immediately then content appears when it is ready.  As part of this work, rewrote the
HTTP caching bit to not use a specific key but allow the remote app dev the ability to
define exactly what should be in the url that is used for caching.  This means user-specific
content can now be returned from macros.  Finally, content retrieval, including web hooks,
are now processed with Java NIO.

Change-Id: Iab8b6df6c1827cabf1a7bab3af3d9434a072f615