ACDEV-2066 Consistenly use UserKey, don't assume add-on users are Crowd-backed
Stash uses ServiceUser to implement add-on users, which are not backed by
Embedded Crowd. Also, UserKey corresponds to a user's ID in Stash to more
gracefully handle user renames.

Unfortunately, ThreeLeggedAuthFilter assumes that the UserKey corresponds to
a user's username in Crowd. As a result, ThreeLeggedAuthFilter fails to
authenticate the add-on user in Stash,

This updates connect to:

* consistently use UserKey instead of a raw String wherever it refers to the
  add-on user's user key
* no longer assume that add-on users are Crowd backed
* replace ConnectAddOnUserService with ConnectUserService and promote it to the
  API to allow the host application to implement the various user access and
  user update functionality that Connect needs.