AC-1704: Ensure encodeQuery uses percent encoding
UrlVariableSubstitutorImpl.encodeQuery now replaces all post-encoding
'+' symbols with '%20'. This is due to the fact that URLEncoder.encode
will encode space as plus. For Query parameters this is legal, however
to stay consistent with previous behavior, we need to percent encode
here.

Also re-add the log-level properties for the commons.httpclient for
Confluence.