AC-1693: Fixed all of the test cases.
There were really subtle errors with Jackson. One was the fact that one of the inner classes was not
private and thus the constructor technically needed four constructor variables instead of three (one
for the parent class itself). And the other is that jackson needs the no argument constructor to be
present for some reason.